+++ History.md
@@ -12,6 +12,7 @@
   * Faster phased restart and worker timeout (#2121)
   * New configuration option to set state file permissions (#2238)
   * `Puma.stats_hash` returns a stats in Hash instead of a JSON string (#2086, #2253)
+  * Add `fork_worker` option and `refork` command for improved copy-on-write performance (#2099)
 
 * Deprecations, Removals and Breaking API Changes
   * `--control` has been removed. Use `--control-url` (#1487)